Complete overhaul. Now there is a run_create_all_figures.py that imports all run functions to create each image from the paper. The images are names 'figure_1', 'supp_figure_N_feature', where N is figure number and feature is what the figure is meant to convey.
Created SensitivityData and TimingData. TimingData should only be used by the person wanting to recreate. It is specific for the machines we used for the paper. SensitivityData can be manipulated. The results should be the same. However, i may add hard copies of the output, and make a run call that passes specific data to the create_figure that is specific for each sensitivity run.